Nu med repository og bedre struktur
