+++ AuctionPlanningServiceAPI/Service/IAuctionPlanningRepository.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using AuctionPlanningServiceAPI.Model;
+using Microsoft.AspNetCore.Mvc;
+
+
+namespace AuctionPlanningServiceAPI.Service
+{
+
+    public interface IAuctionPlanningRepository
+    {
+        public Task<Auction> AddAuction(AuctionDTO auctionDTO);
+
+        public Task<List<Auction>> GetAllAuctions();
+
+        public Task<Auction> DeleteAuction(string id);
+
+        public Task<Auction> GetAuctionByID(string id);
+
+        public Task<Auction> UpdateAuction(string id, AuctionDTO auctionDTO);
+
+    }
+}
